Add adb_rndis USB gadget mode.
Adds a new gadget mode, adb_rndis, that exposes rndis alongside all
interfaces of adb mode (acm, mass_storage, adb).  This mode is entered when
both adbd is running (open on /dev/android_adb_enable) and rndis is enabled
(echo 1 > /sys/devices/virtual/usb_composite/rndis/enable).  When either
adb or rndis is disabled from this mode, it falls back to rndis-only or
adb-only as appropriate.  In addition, synchronization is added to the
gadget's android_enable_function so that userspace activation of these
modes need not be synchronized.

This new mode allows rndis to be used simultaneously with adb and
mass_storage, specifically in Linux where rndis doesn't work (at present)
in stand-alone mode due to misconfigured CDC interface descriptors.
Unfortunately rndis doesn't work in adb mode in Windows as it requires
rndis to appears on the first two interfaces, and the Samsung USB driver
doesn't enumerate rndis in this configuration.  That said, this mode
shouldn't be harmful and retain the previous adb + mass_storage capability.

Change-Id: Ia9cfd8ba13226537dd2f16e0776525392288ccc4