Ignore IOCTL_MFC_BUF_CACHE requests, fixes decoded video artifacts.
Buffer cache support (and this ioctl) is neither implemented in the
crespo 2.6.35 kernel, nor is it ever used in Gingerbread ROMs, including
EL30.  There appears to be partial support for it in this driver, but
it's broken, resulting in artifacts lines in decoded video.  It's most
noticeable on 240p mpeg4 video, but is not limited to it.