Ignore IOCTL_MFC_BUF_CACHE requests, fixes decoded video artifacts.
Buffer cache support (and this ioctl) is neither implemented in the
crespo 2.6.35 kernel, nor is it ever used in Gingerbread ROMs, including
EL30.  There appears to be partial support for it in this driver, but
it's broken, resulting in artifacts lines in decoded video.  It's most
noticeable on 240p mpeg4 video, but is not limited to it.

This also appears to fix the color artifacts at the beginning of 720p
video playback (i.e., those recorded by the camcorder).

Note: When playing back a "large" resolution video (e.g., 1280x720) this
change introduces a "pause" when the Video Player's OSD fades in and out.
This pause does not happen in video players that don't have alpha-blending
effects (e.g., ES File Explorer's built-in player).  The pause also does
not happen on 854x480 resolution videos and smaller.  For resolutions in
between, the extent of the pause/skip varies.

Change-Id: I222bfd604dfe832e37d2ec604555a9b91f549d43