Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I confirm what pawitp says.  From the changelog:

gpu: pvr: Update to DDK 1.8.18.919
- Bump CONFIG_FB_S3C_NR_BUFFERS to 6, to increase the display carveout size by 3MB. This is to make space for video-decode buffers.

Looking into this, all setting CONFIG_FB_S3C_NR_BUFFERS=7 with regard to s3cfb does is make the virtual size of the framebuffer be 480x5600, which seems useless on the surface to me.  I've confirmed that regardless of the NR_BUFFERS setting, Android only uses a double-buffered framebuffer for _most_ GPU stuff.  Thus, itself requires a 480x1600 buffer, or NR_BUFFERS=2.

However, video playback doesn't happen in the "default window" framebuffer but rather in a separate overlay window, the size of which is defined by CONFIG_FB_S3C_NUM_OVLY_WIN=1 CONFIG_FB_S3C_NUM_BUF_OVLY_WIN=3.  Video playback is actually triple buffered, so those are the minimum values.  Now, while these factor into the calculation of FIMD size seprately, it turns out that CONFIG_FB_S3C_NR_BUFFERS=6 is needed at minimum (as the quoted changelog says) for video playback to actually use the overlay window.

I've yet to figure out why that's the case.  Unless a separate contiguous memory space, ~3 MB, is needed for video decoding independent of the overlay space, in which case carving it out of FIMD is easy but unintuitive.  That, or the logic that determines whether to use the overlay window or not is buggy.  I think this is determined in libhwcomposer, but I've not quite figured out the mechanism.

In any event, when video playback doesn't use the overlay window it falls back to the double-buffered framebuffer like everything else.  I don't notice a visual difference between the two, but I haven't looked that closely.

If for some reason we decide to drop the overlay window for video playback, we can turn down those config parameters too, in which case a FIMD size of 3,072,000 B (NR_BUFFERS=2 with no overlay) matches what we use in GB.

Patch-set: 1