Update patch set 3
Patch Set 3:

No, not yet. That requires something I need to take more time looking into. I didn't necessarily get an elaboration into how it's wanted for me to go about doing that without affecting OC, so for now, this is up as a placeholder with 1.2 as the max and 1.3 & 1.1 removed. Only left and required thing to abide by the consensus is the UV. Which is inactive either way. Unless someone prompts to Undervolt with Voltage Control or Setcpu at this point in time, someone using performance settings within CM9 wouldn't have access to it either way.

Patch-set: 3